+++ nodeconductor/structure/migrations/0001_squashed_0021_balancehistory.py
@@ -41,6 +41,7 @@ class Migration(migrations.Migration):
             ],
             options={
                 'abstract': False,
+                'verbose_name': 'organization',
             },
         ),
         migrations.CreateModel(
@@ -48,7 +49,7 @@ class Migration(migrations.Migration):
             fields=[
                 ('id', models.AutoField(verbose_name='ID', serialize=False, auto_created=True, primary_key=True)),
                 ('role_type', models.SmallIntegerField(choices=[(0, 'Owner')])),
-                ('customer', models.ForeignKey(related_name='roles', to='structure.Customer')),
+                ('customer', models.ForeignKey(verbose_name='organization', related_name='roles', to='structure.Customer')),
                 ('permission_group', models.OneToOneField(to='auth.Group')),
             ],
             options={
@@ -64,7 +65,7 @@ class Migration(migrations.Migration):
                 ('created', model_utils.fields.AutoCreatedField(default=django.utils.timezone.now, verbose_name='created', editable=False)),
                 ('modified', model_utils.fields.AutoLastModifiedField(default=django.utils.timezone.now, verbose_name='modified', editable=False)),
                 ('name', models.CharField(max_length=150, verbose_name='name', validators=[nodeconductor.core.validators.validate_name])),
-                ('customer', models.ForeignKey(related_name='projects', on_delete=django.db.models.deletion.PROTECT, to='structure.Customer')),
+                ('customer', models.ForeignKey(verbose_name='organization', related_name='projects', on_delete=django.db.models.deletion.PROTECT, to='structure.Customer')),
             ],
             options={
                 'abstract': False,
@@ -92,7 +93,7 @@ class Migration(migrations.Migration):
                 ('created', model_utils.fields.AutoCreatedField(default=django.utils.timezone.now, verbose_name='created', editable=False)),
                 ('modified', model_utils.fields.AutoLastModifiedField(default=django.utils.timezone.now, verbose_name='modified', editable=False)),
                 ('name', models.CharField(max_length=150, verbose_name='name', validators=[nodeconductor.core.validators.validate_name])),
-                ('customer', models.ForeignKey(related_name='project_groups', on_delete=django.db.models.deletion.PROTECT, to='structure.Customer')),
+                ('customer', models.ForeignKey(verbose_name='organization', related_name='project_groups', on_delete=django.db.models.deletion.PROTECT, to='structure.Customer')),
                 ('projects', models.ManyToManyField(related_name='project_groups', to=b'structure.Project')),
             ],
             options={
@@ -116,7 +117,7 @@ class Migration(migrations.Migration):
             name='ServiceSettings',
             fields=[
                 ('id', models.AutoField(verbose_name='ID', serialize=False, auto_created=True, primary_key=True)),
-                ('customer', models.ForeignKey(related_name='service_settings', blank=True, to='structure.Customer', null=True)),
+                ('customer', models.ForeignKey(verbose_name='organization', related_name='service_settings', blank=True, to='structure.Customer', null=True)),
                 ('name', models.CharField(max_length=150, verbose_name='name', validators=[nodeconductor.core.validators.validate_name])),
                 ('uuid', nodeconductor.core.fields.UUIDField()),
                 ('state', django_fsm.FSMIntegerField(default=1, choices=[(1, 'Sync Scheduled'), (2, 'Syncing'), (3, 'In Sync'), (4, 'Erred')])),
@@ -142,7 +143,7 @@ class Migration(migrations.Migration):
                 ('id', models.AutoField(verbose_name='ID', serialize=False, auto_created=True, primary_key=True)),
                 ('created', model_utils.fields.AutoCreatedField(default=django.utils.timezone.now, editable=False)),
                 ('amount', models.DecimalField(max_digits=9, decimal_places=3)),
-                ('customer', models.ForeignKey(to='structure.Customer')),
+                ('customer', models.ForeignKey(verbose_name='organization', to='structure.Customer')),
             ],
         ),
     ]