+++ nodeconductor/structure/tests/test_customer.py
@@ -231,7 +231,7 @@ class CustomerApiManipulationTest(UrlResolverMixin, test.APISimpleTestCase):
         response = self.client.delete(self._get_customer_url(self.fixture.customer))
 
         self.assertEqual(response.status_code, status.HTTP_409_CONFLICT)
-        self.assertDictContainsSubset({'detail': 'Cannot delete customer with existing projects'},
+        self.assertDictContainsSubset({'detail': 'Cannot delete organization with existing projects'},
                                       response.data)
 
     # Creation tests
@@ -440,21 +440,18 @@ class CustomerUnicodeTest(TransactionTestCase):
 
 @ddt
 class CustomerUsersListTest(test.APITransactionTestCase):
+    all_users = ('staff', 'owner', 'manager', 'admin', 'customer_support', 'project_support', 'global_support')
+
     def setUp(self):
         self.fixture = fixtures.ProjectFixture()
-        self.staff = self.fixture.staff
-        self.owner = self.fixture.owner
-        self.admin = self.fixture.admin
-        self.manager = self.fixture.manager
-        self.customer = self.fixture.customer
-        self.customer_support = self.fixture.customer_support
-        self.project_support = self.fixture.project_support
-        self.global_support = self.fixture.global_support
-        self.url = factories.CustomerFactory.get_url(self.customer, action='users')
+        self.url = factories.CustomerFactory.get_url(self.fixture.customer, action='users')
 
-    @data('staff', 'owner', 'manager', 'admin', 'customer_support', 'project_support', 'global_support')
+    @data(*all_users)
     def test_user_can_list_customer_users(self, user):
-        self.client.force_authenticate(getattr(self, user))
+        self.client.force_authenticate(getattr(self.fixture, user))
+        # call fixture to initiate all users:
+        for user in self.all_users:
+            getattr(self.fixture, user)
         response = self.client.get(self.url)
 
         self.assertEqual(response.status_code, status.HTTP_200_OK)
@@ -462,8 +459,8 @@ class CustomerUsersListTest(test.APITransactionTestCase):
 
         self.assertSetEqual({user['role'] for user in response.data}, {'owner', 'support', None, None})
         self.assertSetEqual({user['uuid'] for user in response.data},
-                            {self.owner.uuid.hex, self.admin.uuid.hex, self.manager.uuid.hex,
-                             self.customer_support.uuid.hex, self.project_support.uuid.hex})
+                            {self.fixture.owner.uuid.hex, self.fixture.admin.uuid.hex, self.fixture.manager.uuid.hex,
+                             self.fixture.customer_support.uuid.hex, self.fixture.project_support.uuid.hex})
         self.assertSetEqual({user['projects'] and user['projects'][0]['role'] or None
                              for user in response.data}, {None, 'admin', 'manager', 'support'})
 
@@ -471,6 +468,25 @@ class CustomerUsersListTest(test.APITransactionTestCase):
         self.client.force_authenticate(self.fixture.user)
         response = self.client.get(self.url)
         self.assertEqual(response.status_code, status.HTTP_404_NOT_FOUND)
+
+    def test_users_ordering_by_concatenated_name(self):
+        walter = factories.UserFactory(full_name='', username='walter')
+        admin = factories.UserFactory(full_name='admin', username='zzz')
+        alice = factories.UserFactory(full_name='', username='alice')
+        dave = factories.UserFactory(full_name='dave', username='dave')
+        expected_order = [admin, alice, dave, walter]
+        for user in expected_order:
+            self.fixture.customer.add_user(user, CustomerRole.OWNER)
+
+        self.client.force_authenticate(self.fixture.staff)
+        response = self.client.get(self.url + '?o=concatenated_name')
+        for serialized_user, expected_user in zip(response.data, expected_order):
+            self.assertEqual(serialized_user['uuid'], expected_user.uuid.hex)
+
+        # reversed order
+        response = self.client.get(self.url + '?o=-concatenated_name')
+        for serialized_user, expected_user in zip(response.data, expected_order[::-1]):
+            self.assertEqual(serialized_user['uuid'], expected_user.uuid.hex)
 
 
 @ddt